[REFACTOR][PYTHON] Move tvm.contrib.tar to tvm.support.tar
tvm.contrib.tar provides tar()/untar() helpers used by runtime/module.py's
export_library flow, rpc/server.py, and the meta-schedule local builder.
It is part of the same library-packaging pipeline as cc.py and utils.py.
Moving it to tvm.support completes the set of archive/filesystem helpers
under the support namespace.
