+++ python/tvm/s_tir/meta_schedule/builder/local_builder.py
@@ -280,7 +280,7 @@ def default_export(mod: Module) -> str:
     artifact_path : str
         The path to the exported Module.
     """
-    from tvm.contrib.tar import tar  # pylint: disable=import-outside-toplevel
+    from tvm.support.tar import tar  # pylint: disable=import-outside-toplevel
 
     artifact_path = os.path.join(tempfile.mkdtemp(), "tvm_tmp_mod." + tar.output_format)
     mod.export_library(artifact_path, fcompile=tar)
